+++ src/main/java/dev/willyelton/variable_redstone/VariableRedstone.java
@@ -1,8 +1,10 @@
 package dev.willyelton.variable_redstone;
 
+import dev.willyelton.variable_redstone.common.ChunkPulseLengthData;
+import dev.willyelton.variable_redstone.common.VariableRedstoneConfig;
 import dev.willyelton.variable_redstone.common.block.VariableButton;
-import dev.willyelton.variable_redstone.common.block.VariableRedstoneBlock;
 import dev.willyelton.variable_redstone.common.block.VariableLever;
+import dev.willyelton.variable_redstone.common.block.VariableRedstoneBlock;
 import dev.willyelton.variable_redstone.common.block.VariableRedstoneTorch;
 import dev.willyelton.variable_redstone.common.block.VariableRedstoneWallTorch;
 import net.minecraft.core.Direction;
@@ -19,8 +21,13 @@ import net.minecraft.world.level.block.state.properties.IntegerProperty;
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.fml.ModContainer;
 import net.neoforged.fml.common.Mod;
+import net.neoforged.fml.config.ModConfig;
+import net.neoforged.neoforge.attachment.AttachmentType;
 import net.neoforged.neoforge.registries.DeferredHolder;
 import net.neoforged.neoforge.registries.DeferredRegister;
+import net.neoforged.neoforge.registries.NeoForgeRegistries;
+
+import java.util.function.Supplier;
 
 @Mod(VariableRedstone.MODID)
 public class VariableRedstone {
@@ -29,6 +36,7 @@ public class VariableRedstone {
     public static final DeferredRegister.Blocks BLOCKS = DeferredRegister.createBlocks(MODID);
     public static final DeferredRegister.Items ITEMS = DeferredRegister.createItems(MODID);
     public static final DeferredRegister<CreativeModeTab> CREATIVE_MODE_TABS = DeferredRegister.create(Registries.CREATIVE_MODE_TAB, MODID);
+    public static final DeferredRegister<AttachmentType<?>> ATTACHMENT_TYPES = DeferredRegister.create(NeoForgeRegistries.ATTACHMENT_TYPES, MODID);
 
     // Blocks
     public static final DeferredHolder<Block, Block> VARIABLE_REDSTONE_BLOCK = BLOCKS.registerBlock("variable_redstone_block", VariableRedstoneBlock::new);
@@ -43,6 +51,13 @@ public class VariableRedstone {
 
     public static final DeferredHolder<Block, VariableButton> VARIABLE_BUTTON = BLOCKS.register("variable_button", VariableButton::new);
     public static final DeferredHolder<Item, BlockItem> VARIABLE_BUTTON_BLOCK_ITEM = ITEMS.registerSimpleBlockItem(VARIABLE_BUTTON);
+
+    // Data Attachments
+    public static final Supplier<AttachmentType<ChunkPulseLengthData>> CHUNK_PULSE_LENGTH = ATTACHMENT_TYPES.register("chunk_pulse_length",
+            () -> AttachmentType.builder(ChunkPulseLengthData::new)
+                    .serialize(ChunkPulseLengthData.CODEC)
+                    .sync(ChunkPulseLengthData.STREAM_CODEC)
+                    .build());
 
     // Creative Tab
     public static DeferredHolder<CreativeModeTab, CreativeModeTab> TAB = CREATIVE_MODE_TABS.register("variable_redstone_tab", () ->
@@ -60,7 +75,10 @@ public class VariableRedstone {
     public VariableRedstone(IEventBus modEventBus, ModContainer modContainer) {
         BLOCKS.register(modEventBus);
         ITEMS.register(modEventBus);
+        ATTACHMENT_TYPES.register(modEventBus);
         CREATIVE_MODE_TABS.register(modEventBus);
+
+        modContainer.registerConfig(ModConfig.Type.SERVER, VariableRedstoneConfig.SERVER_CONFIG);
     }
 
     public static ResourceLocation rl(String path) {